python/requirements.txt
aiohttp
aioprocessing
flashbots
multicall==0.7.7
nest-asyncio
numpy
pandas
tqdm
web3==6.11
websockets
loguru
python-dotenv
eth-abi
# run pip install --upgrade websockets after completion
